feat(cards): implement card moving between columns
Add MoveCardCommand and handler to support moving cards between columns with position recalculation. Update Card and BoardList entities to handle new Position type and implement placement logic. Modify DTOs and database configuration to support the new feature.

Remove unused keyboard event handlers from frontend components and update modal close event.

+++ backend/Organetto/Organetto.UseCases/Boards/Data/CardDto.cs
@@ -1,6 +1,6 @@
 ﻿namespace Organetto.UseCases.Boards.Data
 {
-    public record CardDto(long Id, string Title, string Description, int Position, DateTimeOffset? DueDate)
+    public record CardDto(long Id, string Title, string Description, long Position, DateTimeOffset? DueDate)
     {
         public CardDto() : this(0, string.Empty, string.Empty, 0, null)
         {
